requirements.txt
astroid==2.1.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
Django==2.1.4
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-fontawesome==1.0
django-froala-editor==2.9.1.post2
django-microsoft-auth==1.1.1
django-precise-bbcode==1.2.10
django-tinymce==2.7.0
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysql-connector-python==8.0.13
oauthlib==2.1.0
Pillow==5.3.0
protobuf==3.6.1
pylint==2.2.2
PyMySQL==0.9.3
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
requests==2.21.0
requests-oauthlib==1.0.0
six==1.12.0
sqlparse==0.2.4
urllib3==1.24.1
wrapt==1.10.11
